Update headers to match so_model_module_simple.cpp

+++ nw4r/include/nw4r/g3d/g3d_resmdl.h
@@ -21,7 +21,7 @@ namespace nw4r {
             ResMat GetResMat(const char* name) const;
             u32 GetResNodeNumEntries();
             ResNode GetResNode(const int nodeIndex) const;
-            ResNode GetResNode(const u64 nodeIndex) const;
+            ResNode GetResNode(const u32 nodeIndex) const;
             ResNode GetResNode(const char* nodeName) const;
         };
